+++ pkg/store/storepb/testutil/series.go
@@ -16,15 +16,17 @@ import (
 	"time"
 
 	"github.com/cespare/xxhash"
-
+	"github.com/efficientgo/core/testutil"
 	"github.com/gogo/protobuf/types"
+	"github.com/prometheus/prometheus/model/histogram"
 	"github.com/prometheus/prometheus/model/labels"
+	"github.com/prometheus/prometheus/storage"
 	"github.com/prometheus/prometheus/tsdb"
+	"github.com/prometheus/prometheus/tsdb/chunkenc"
 	"github.com/prometheus/prometheus/tsdb/chunks"
 	"github.com/prometheus/prometheus/tsdb/index"
 	"github.com/prometheus/prometheus/tsdb/wlog"
-
-	"github.com/efficientgo/core/testutil"
+	"go.uber.org/atomic"
 
 	"github.com/thanos-io/thanos/pkg/store/hintspb"
 	"github.com/thanos-io/thanos/pkg/store/labelpb"
@@ -51,6 +53,7 @@ type HeadGenOptions struct {
 	WithWAL       bool
 	PrependLabels labels.Labels
 	SkipChunks    bool // Skips chunks in returned slice (not in generated head!).
+	SampleType    chunkenc.ValueType
 
 	Random *rand.Rand
 }
@@ -65,6 +68,10 @@ func CreateHeadWithSeries(t testing.TB, j int, opts HeadGenOptions) (*tsdb.Head,
 	}
 	if opts.ScrapeInterval == 0 {
 		opts.ScrapeInterval = 1 * time.Millisecond
+	}
+	// Use float type if sample type is not set.
+	if opts.SampleType == chunkenc.ValNone {
+		opts.SampleType = chunkenc.ValFloat
 	}
 
 	fmt.Printf(
@@ -86,27 +93,26 @@ func CreateHeadWithSeries(t testing.TB, j int, opts HeadGenOptions) (*tsdb.Head,
 
 	headOpts := tsdb.DefaultHeadOptions()
 	headOpts.ChunkDirRoot = opts.TSDBDir
+	headOpts.EnableNativeHistograms = *atomic.NewBool(true)
 	h, err := tsdb.NewHead(nil, nil, w, nil, headOpts, nil)
 	testutil.Ok(t, err)
 
 	app := h.Appender(context.Background())
 	for i := 0; i < opts.Series; i++ {
 		tsLabel := j*opts.Series*opts.SamplesPerSeries + i*opts.SamplesPerSeries
-		ref, err := app.Append(
-			0,
-			labels.FromStrings("foo", "bar", "i", fmt.Sprintf("%07d%s", tsLabel, LabelLongSuffix)),
-			int64(tsLabel)*opts.ScrapeInterval.Milliseconds(),
-			opts.Random.Float64(),
-		)
-		testutil.Ok(t, err)
-
-		for is := 1; is < opts.SamplesPerSeries; is++ {
-			_, err := app.Append(ref, nil, int64(tsLabel+is)*opts.ScrapeInterval.Milliseconds(), opts.Random.Float64())
-			testutil.Ok(t, err)
+		switch opts.SampleType {
+		case chunkenc.ValFloat:
+			appendFloatSamples(t, app, tsLabel, opts)
+		case chunkenc.ValHistogram:
+			appendHistogramSamples(t, app, tsLabel, opts)
 		}
 	}
 	testutil.Ok(t, app.Commit())
 
+	return h, ReadSeriesFromBlock(t, h, opts.PrependLabels, opts.SkipChunks)
+}
+
+func ReadSeriesFromBlock(t testing.TB, h tsdb.BlockReader, extLabels labels.Labels, skipChunks bool) []*storepb.Series {
 	// Use TSDB and get all series for assertion.
 	chks, err := h.Chunks()
 	testutil.Ok(t, err)
@@ -119,7 +125,7 @@ func CreateHeadWithSeries(t testing.TB, j int, opts HeadGenOptions) (*tsdb.Head,
 	var (
 		lset       labels.Labels
 		chunkMetas []chunks.Meta
-		expected   = make([]*storepb.Series, 0, opts.Series)
+		expected   = make([]*storepb.Series, 0)
 	)
 
 	var builder labels.ScratchBuilder
@@ -128,9 +134,9 @@ func CreateHeadWithSeries(t testing.TB, j int, opts HeadGenOptions) (*tsdb.Head,
 	for all.Next() {
 		testutil.Ok(t, ir.Series(all.At(), &builder, &chunkMetas))
 		lset = builder.Labels()
-		expected = append(expected, &storepb.Series{Labels: labelpb.ZLabelsFromPromLabels(append(opts.PrependLabels.Copy(), lset...))})
+		expected = append(expected, &storepb.Series{Labels: labelpb.ZLabelsFromPromLabels(append(extLabels.Copy(), lset...))})
 
-		if opts.SkipChunks {
+		if skipChunks {
 			continue
 		}
 
@@ -146,12 +152,60 @@ func CreateHeadWithSeries(t testing.TB, j int, opts HeadGenOptions) (*tsdb.Head,
 			expected[len(expected)-1].Chunks = append(expected[len(expected)-1].Chunks, storepb.AggrChunk{
 				MinTime: c.MinTime,
 				MaxTime: c.MaxTime,
-				Raw:     &storepb.Chunk{Type: storepb.Chunk_XOR, Data: chEnc.Bytes(), Hash: xxhash.Sum64(chEnc.Bytes())},
+				Raw: &storepb.Chunk{
+					Data: chEnc.Bytes(),
+					Type: storepb.Chunk_Encoding(chEnc.Encoding() - 1),
+					Hash: xxhash.Sum64(chEnc.Bytes()),
+				},
 			})
 		}
 	}
 	testutil.Ok(t, all.Err())
-	return h, expected
+	return expected
+}
+
+func appendFloatSamples(t testing.TB, app storage.Appender, tsLabel int, opts HeadGenOptions) {
+	ref, err := app.Append(
+		0,
+		labels.FromStrings("foo", "bar", "i", fmt.Sprintf("%07d%s", tsLabel, LabelLongSuffix)),
+		int64(tsLabel)*opts.ScrapeInterval.Milliseconds(),
+		opts.Random.Float64(),
+	)
+	testutil.Ok(t, err)
+
+	for is := 1; is < opts.SamplesPerSeries; is++ {
+		_, err := app.Append(ref, nil, int64(tsLabel+is)*opts.ScrapeInterval.Milliseconds(), opts.Random.Float64())
+		testutil.Ok(t, err)
+	}
+}
+
+func appendHistogramSamples(t testing.TB, app storage.Appender, tsLabel int, opts HeadGenOptions) {
+	sample := &histogram.Histogram{
+		Schema:        0,
+		Count:         9,
+		Sum:           -3.1415,
+		ZeroCount:     12,
+		ZeroThreshold: 0.001,
+		NegativeSpans: []histogram.Span{
+			{Offset: 0, Length: 4},
+			{Offset: 1, Length: 1},
+		},
+		NegativeBuckets: []int64{1, 2, -2, 1, -1},
+	}
+
+	ref, err := app.AppendHistogram(
+		0,
+		labels.FromStrings("foo", "bar", "i", fmt.Sprintf("%07d%s", tsLabel, LabelLongSuffix)),
+		int64(tsLabel)*opts.ScrapeInterval.Milliseconds(),
+		sample,
+		nil,
+	)
+	testutil.Ok(t, err)
+
+	for is := 1; is < opts.SamplesPerSeries; is++ {
+		_, err := app.AppendHistogram(ref, nil, int64(tsLabel+is)*opts.ScrapeInterval.Milliseconds(), sample, nil)
+		testutil.Ok(t, err)
+	}
 }
 
 // SeriesServer is test gRPC storeAPI series server.
@@ -265,7 +319,9 @@ func TestServerSeries(t testutil.TB, store storepb.StoreServer, cases ...*Series
 								if len(c.ExpectedSeries[j].Chunks) > 20 {
 									testutil.Equals(t, len(c.ExpectedSeries[j].Chunks), len(srv.SeriesSet[j].Chunks), "%v series chunks number mismatch", j)
 								}
-								testutil.Equals(t, c.ExpectedSeries[j].Chunks, srv.SeriesSet[j].Chunks, "%v series chunks mismatch", j)
+								for ci := range c.ExpectedSeries[j].Chunks {
+									testutil.Equals(t, c.ExpectedSeries[j].Chunks[ci], srv.SeriesSet[j].Chunks[ci], "%v series chunks mismatch %v", j, ci)
+								}
 							}
 						}
 					} else {